Added notification mechanism
We added a format command for fast formatting with prettier
since it will increase consistency.

Added a new api module for handling requests in a more
consistent way. This module has a core method `handleApiRequest`
intercepting requests and throws an error when status is not 200.
This a crucial step for our notification system.

We also integrated our notification system with forms. It makes
it easier to understand the problem when an error returned from the
backend.